refactor: DevBalanceRepository 책임 분리
